Expose referer token property as its required by some users for token generation
A 10.2.2 server that has been secured returns the following error when generating a token if the referer is not provided.

---
Exception in generating tokenFailed to generate a token for the user.Client ID is required for long expiration tokens
---

I have taken the approach of exposing the referer property through the creation of the gateway (as this is where the username + password was being set) as it seems impossible to set otherwise.